backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Async & Task Queue
celery==5.3.4
redis==5.0.1
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
elasticsearch==8.10.0

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1
pdf2image==1.16.3

# OCR & Document Processing
surya-ocr==0.4.14
python-mrz==0.6.3
pytesseract==0.3.10

# Face Recognition
insightface==0.7.3
onnxruntime==1.16.3
scikit-learn==1.3.2

# ML/AI
torch==2.0.1
torchvision==0.15.2

# Utilities
numpy==1.26.0
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1

# Cryptography
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
